Leetcode questions/LC_1431/src/main/java/org/akhil/Main.java
package org.akhil;

import java.util.ArrayList;
import java.util.List;

public class Main {
    public static void main(String[] args) {
        int[] candies = {2,3,5,1,3};
        int extraCandies = 3;

        System.out.println(kidsWithCandies(candies, extraCandies));
    }

    public static List<Boolean> kidsWithCandies(int[] candies, int extraCandies) {
        List<Boolean> result = new ArrayList<>();
        int max = Integer.MIN_VALUE;
        for(int number:candies){
            if(max<number)
                max = number;
        }
        for(int i=0; i<candies.length; i++){
            result.add(max<=(candies[i]+extraCandies));
        }
        return result;
    }
}